New extractor test for only builtin data types
A new test has been added to the test factory to make sure the extractor outputs only builtin data types or np.nan. The extractors were also updated to pass this test

+++ panoptes_aggregation/extractors/utilities.py
@@ -149,6 +149,6 @@ def get_feedback_info(feedback_dict):
 
     # calculate an agreement score for this classification
     # (ratio of successful classifications to number of gold standards)
-    feedback_data['agreement_score'] = np.mean(feedback_data['success'])
+    feedback_data['agreement_score'] = float(np.mean(feedback_data['success']))
 
     return feedback_data
